Rename project and middleware
