+++ packages/extension-sample/src/types/config.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import type {ApplicationExtensionConfig} from '@salesforce/pwa-kit-extension-support/types'
+import type {ApplicationExtensionConfig} from '@salesforce/pwa-kit-application-extensibility/types'
 
 // This is where you are going to define the configuration type for your App Extension. This is used in the constructor
 // of the extension itself. Update this config type to your specific needs!
