+++ packages/pwa-kit-application-extensibility/.eslintignore
