+++ packages/pwa-kit-application-extensibility/.eslintrc.js
