+++ packages/pwa-kit-application-extensibility/.gitignore
