+++ packages/pwa-kit-application-extensibility/.npmignore
