+++ packages/pwa-kit-application-extensibility/.prettierignore
