+++ packages/pwa-kit-application-extensibility/.prettierrc.yaml
