+++ packages/pwa-kit-application-extensibility/LICENSE
