+++ packages/pwa-kit-application-extensibility/babel.config.js
