+++ packages/pwa-kit-application-extensibility/jest.config.js
