+++ packages/pwa-kit-application-extensibility/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@salesforce/pwa-kit-extension-support",
+  "name": "@salesforce/pwa-kit-application-extensibility",
   "version": "4.0.0-dev",
   "description": "",
   "homepage": "https://github.com/SalesforceCommerceCloud/pwa-kit#readme",
