+++ packages/pwa-kit-application-extensibility/src/configs/babel/plugin-application-extensions.ts
