+++ packages/pwa-kit-application-extensibility/src/configs/utils.ts
