+++ packages/pwa-kit-application-extensibility/src/configs/webpack/index.ts
