+++ packages/pwa-kit-application-extensibility/src/configs/webpack/loaders/extensions-loader.ts
