+++ packages/pwa-kit-application-extensibility/src/configs/webpack/plugins/index.ts
