+++ packages/pwa-kit-application-extensibility/src/configs/webpack/plugins/overrides-resolver-plugin.ts
