+++ packages/pwa-kit-application-extensibility/src/configs/webpack/types.ts
