+++ packages/pwa-kit-application-extensibility/src/express/assets/application-extensions-placeholder.ts
