+++ packages/pwa-kit-application-extensibility/src/express/index.ts
@@ -12,7 +12,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 import './application-extension'
-import './application-extensions-middleware'
+import './apply-application-extensions'
 
 export * from './application-extension'
-export {default as applicationExtensionMiddleware} from './application-extensions-middleware'
+export {default as applyApplicationExtensions} from './apply-application-extensions'
