+++ packages/pwa-kit-application-extensibility/src/index.ts
