+++ packages/pwa-kit-application-extensibility/src/react/ApplicationExtension.ts
