+++ packages/pwa-kit-application-extensibility/src/react/assets/application-extensions-placeholder.ts
