+++ packages/pwa-kit-application-extensibility/src/react/components/index.ts
