+++ packages/pwa-kit-application-extensibility/src/react/components/withApplicationExtensions.tsx
