+++ packages/pwa-kit-application-extensibility/src/react/contexts/ApplicationExtensionsContext.tsx
