+++ packages/pwa-kit-application-extensibility/src/react/contexts/ApplicationExtensionsProvider.tsx
