+++ packages/pwa-kit-application-extensibility/src/react/hooks/index.ts
