+++ packages/pwa-kit-application-extensibility/src/react/hooks/useApplicationExtensions.tsx
