+++ packages/pwa-kit-application-extensibility/src/react/index.ts
