+++ packages/pwa-kit-application-extensibility/src/react/utils/index.ts
