+++ packages/pwa-kit-application-extensibility/src/shared/utils/extensibility-utils.test.js
