+++ packages/pwa-kit-application-extensibility/src/shared/utils/index.ts
