+++ packages/pwa-kit-application-extensibility/src/shared/utils/resolver-utils.test.js
