+++ packages/pwa-kit-application-extensibility/src/shared/utils/universal-utils.ts
