+++ packages/pwa-kit-application-extensibility/src/types.ts
