+++ packages/pwa-kit-application-extensibility/tsconfig.json
