+++ packages/pwa-kit-runtime/package.json
@@ -47,7 +47,7 @@
   "devDependencies": {
     "@loadable/component": "^5.15.3",
     "@salesforce/pwa-kit-dev": "4.0.0-dev",
-    "@salesforce/pwa-kit-extension-support": "4.0.0-dev",
+    "@salesforce/pwa-kit-application-extensibility": "4.0.0-dev",
     "@serverless/event-mocks": "^1.1.1",
     "@types/express": "^4.17.21",
     "aws-lambda-mock-context": "^3.2.1",
